backend/auth-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
pymongo
passlib
python-jose
python-dotenv
email-validator
python-multipart
